chore: remove ExpirationDateTimePicker
