+++ packages/shared/src/components/inputs/index.js
@@ -6,7 +6,6 @@ export { default as ClosableInput } from './ClosableInput.svelte'
 export { default as Dropdown } from './Dropdown.svelte'
 export { default as Dropdown2 } from './Dropdown2.svelte'
 export { default as Dropzone } from './Dropzone.svelte'
-export { default as ExpirationDateTimePicker } from './ExpirationDateTimePicker.svelte'
 export { default as IconInput } from './IconInput.svelte'
 export { default as ImportTextfield } from './ImportTextfield.svelte'
 export { default as Input } from './Input.svelte'
